Add bindings for wlr/util/log.h
wlr-log takes a Scheme format string and arguments, rather than a
C-style format string. It is defined as a macro so that it can use
get-line-number to better mimic regular wlr_log output.

wlr-log-init takes an optional argument for a custom logger function,
but such a function must be written in C. Supporting Scheme loggers
would require that all wlroots functions are defined with
foreign-safe-lambda, which would cause a minor garbage collection at
every call site.