build: use .egg file instead of makefile
Use the egg description file to describe the build process rather than a
makefile.

The key challenge to using the .egg file is getting it to use
pkg-config. This is accomplished by using a custom build script which
calls pkg-config and prepends the arguments to the command line before
invoking csc. This is wasteful, since we are calling pkg-config twice
for each object rather than twice for the whole build.

On the upside, the egg now supports linking with the -static option to
csc, and it doesn't produce warnings when importing (wlr ...) modules.

The makefile is now relegated to building the example programs, and
providing a 'clean' target to get rid of the mess that chicken-install
creates.